refactor: 더 넓은 범위의 숫자 입력을 받기 위해 Int형에서 Long형으로 리팩토링

+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,5 +1,5 @@
 package calculator
 
 interface Calculator {
-    fun execute(numbers: MutableList<Int>): Int
+    fun execute(numbers: MutableList<Long>): Long
 }
