+++ src/main/kotlin/calculator/Validator.kt
@@ -4,7 +4,7 @@ class Validator {
 
     fun validatePositiveNumber(number: String): Boolean {
         return try {
-            if (number.toInt() >= 0) {
+            if (number.toLong() >= 0) {
                 true
             } else {
                 throw IllegalArgumentException("입력하신 숫자는 양수가 아닙니다.")
